cli: Move the {read,write}OrtResult() functions to Extensions.kt
Because they are extension functions.
