+++ cli/src/main/kotlin/utils/Extensions.kt
@@ -19,7 +19,21 @@
 
 package org.ossreviewtoolkit.cli.utils
 
+import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.GroupableOption
+
+import java.io.File
+
+import kotlin.time.measureTime
+import kotlin.time.measureTimedValue
+
+import org.ossreviewtoolkit.model.HashAlgorithm
+import org.ossreviewtoolkit.model.OrtResult
+import org.ossreviewtoolkit.model.readValue
+import org.ossreviewtoolkit.model.writeValue
+import org.ossreviewtoolkit.utils.formatSizeInMib
+import org.ossreviewtoolkit.utils.log
+import org.ossreviewtoolkit.utils.perf
 
 fun <T : GroupableOption> T.group(name: String): T = apply { groupName = name }
 
@@ -28,3 +42,35 @@ fun <T : GroupableOption> T.inputGroup(): T = group(OPTION_GROUP_INPUT)
 fun <T : GroupableOption> T.outputGroup(): T = group(OPTION_GROUP_OUTPUT)
 
 fun <T : GroupableOption> T.configurationGroup(): T = group(OPTION_GROUP_CONFIGURATION)
+
+/**
+ * Read [ortFile] into an [OrtResult] and return it. Make sure that information about project scopes is available
+ * (by calling [OrtResult.withResolvedScopes]), so that it can be processed.
+ */
+fun CliktCommand.readOrtResult(ortFile: File): OrtResult {
+    log.debug { "Input ORT result file has SHA-1 hash ${HashAlgorithm.SHA1.calculate(ortFile)}." }
+
+    val (ortResult, duration) = measureTimedValue { ortFile.readValue<OrtResult>().withResolvedScopes() }
+
+    log.perf {
+        "Read ORT result from '${ortFile.name}' (${ortFile.formatSizeInMib}) in ${duration.inWholeMilliseconds}ms."
+    }
+
+    return ortResult
+}
+
+/**
+ * Write the [ortResult] to all [outputFiles] for the given [resultName].
+ */
+fun CliktCommand.writeOrtResult(ortResult: OrtResult, outputFiles: Collection<File>, resultName: String) {
+    outputFiles.forEach { file ->
+        println("Writing $resultName result to '$file'.")
+        val duration = measureTime { file.writeValue(ortResult) }
+
+        log.perf {
+            "Wrote ORT result to '${file.name}' (${file.formatSizeInMib}) in ${duration.inWholeMilliseconds}ms."
+        }
+
+        log.debug { "Output ORT result file has SHA-1 hash ${HashAlgorithm.SHA1.calculate(file)}." }
+    }
+}
